requirements.txt
cfgrib                      # Reading Gribs.
cython                      # C-Extensions.
dask                        # XArray Dependency.
imutils                     # Utilities for OpenCV.
lxml                        # XML writing.
matplotlib                  # Plotting.
pyproj                      # Geospatial Calculations.
numpy                       # Math.
opencv-contrib-python       # OpenCV in Python.
shapely                     # Geometry.
toolz                       # XArray Dependency.
xarray                      # Better pandas.
